+++ lib/anise/annotative.rb
@@ -1,0 +1,7 @@
+module Anise
+
+  module Annotative
+
+  end
+
+end